Bluetooth : RFCOMM Frame size change
RFCOMM Frame size changes to accomodate the credits in the data packet.
Earlier implementation was not considering the 1 Byte required for
Credits information in data packet and hence was considering the
RFCOMM header plus FCS as 5 Bytes.
Because of this if the remote device sends a data packet with
negotiated frame size and credits information,packet size exceeds L2CAPMTU
and will be discarded by L2CAP.

CRs-fixed: 303518

Signed-off-by: Mallikarjuna GB <gbmalli@codeaurora.org>
(cherry picked from commit 9e11a74)

Change-Id: I9c80c2b292fc5f7b808192ccbe198c0dcae497b6
Signed-off-by: Oguri, Srivalli <oguri@codeaurora.org>